식재료 상세 Request, Response DTO 통합
