+++ src/main/kotlin/mara/server/domain/ingredient/IngredientDetailDto.kt
@@ -2,6 +2,15 @@ package mara.server.domain.ingredient
 
 import java.time.LocalDateTime
 
+data class IngredientDetailRequest(
+    val quantity: Int,
+    val location: String,
+    val memo: String,
+    val addDate: LocalDateTime,
+    val expirationDate: LocalDateTime,
+    val isDeleted: Boolean? = null
+)
+
 data class IngredientDetailResponse(
     val ingredientDetailId: Long,
     val quantity: Int,
